Mudança de tipo de dados das colunas de valores
Pq o double não estava sendo encontrado, então mudamos para Number; Além disso, o boolean não existe, então mudamos para char

+++ sql/tables/Cosmetico.sql
@@ -4,7 +4,7 @@ CREATE TABLE Cosmetico(
 	TipoProduto CHAR(100),
 	Fabricante VARCHAR (100),
 	Descricao VARCHAR(255),
-	Valor DOUBLE,
+	Valor NUMBER(10,3),
 	Peso INT,
 	Validade DATE,
 	Marca CHAR(255),
